move pauli word is_commuting to qml.is_commuting
